+++ cmd/trigger-dictionary-check/.infra/config.json
