+++ cmd/trigger-dictionary-check/.infra/readme.md
