+++ cmd/trigger-dictionary-check/main.go
@@ -1,0 +1,49 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"os"
+	"runtime/debug"
+
+	"github.com/Mad-Pixels/applingo-api/pkg/serializer"
+	"github.com/Mad-Pixels/applingo-api/pkg/trigger"
+
+	"github.com/aws/aws-lambda-go/events"
+	"github.com/aws/aws-lambda-go/lambda"
+	"github.com/aws/aws-sdk-go-v2/config"
+	"github.com/pkg/errors"
+	"github.com/rs/zerolog"
+)
+
+var (
+	awsRegion = os.Getenv("AWS_REGION")
+)
+
+func init() {
+	debug.SetGCPercent(500)
+
+	_, err := config.LoadDefaultConfig(context.Background(), config.WithRegion(awsRegion))
+	if err != nil {
+		panic("unable to load AWS SDK config: " + err.Error())
+	}
+}
+
+func handler(ctx context.Context, log zerolog.Logger, record json.RawMessage) error {
+	var dynamoDBEvent events.DynamoDBEventRecord
+	if err := serializer.UnmarshalJSON(record, &dynamoDBEvent); err != nil {
+		return errors.Wrap(err, "failed to unmarshal DynamoDB event")
+	}
+	log.Error().Any("event", dynamoDBEvent).Msg("event")
+
+	return nil
+}
+
+func main() {
+	lambda.Start(
+		trigger.NewLambda(
+			trigger.Config{MaxWorkers: 6},
+			handler,
+		).Handle,
+	)
+}
