--- cmd/trigger-sqs-to-job-put-csv/.infra/config.json
@@ -1,8 +1,0 @@
-{
-  "memory_size": 128,
-  "timeout": 4,
-  "envs": {
-    "SERVICE_DICTIONARY_BUCKET": "${dictionary_bucket_name}",
-    "SERVICE_PROCESSING_BUCKET": "${processing_bucket_name}"
-  }
-}
