--- cmd/trigger-sqs-to-job-put-csv/.infra/readme.md
@@ -1,3 +1,0 @@
-# Description
-
-Lambda for processing dictionary file.
